+++ src/player.h
@@ -1546,7 +1546,7 @@ class player : public Character
         int blocks_left;
         int stim;
         int radiation;
-        signed long cash;
+        int cash;
         int movecounter;
         std::shared_ptr<monster> mounted_creature;
         bool death_drops;// Turned to false for simulating NPCs on distant missions so they don't drop all their gear in sight
